Make initial population dumber, and do more iterations to compensate
Dumber: remove constructive heuristics, and instead create purely random individuals. This does not reduce the quality of the solutions according to the benchmark script, but it does significantly simplify the codebase.

More iterations: local search is *slow*, and there are instances where we never got beyond the construction of the initial population. These changes remedy that, by making initial construction really fast.