Misc. improvements/simplifications (#15)
* Make initial population dumber, and do more iterations to compensate

Dumber: remove constructive heuristics, and instead create purely random individuals. This does not reduce the quality of the solutions according to the benchmark script, but it does significantly simplify the codebase.

More iterations: local search is *slow*, and there are instances where we never got beyond the construction of the initial population. These changes remedy that, by making initial construction really fast.

* Move indivsPerProximity management into Individual

* Simplify population management

* Remove broken SREX crossover, add configuration option for nbCrossovers

* Clean-up some of LocalSearch's implementation

* Simplify matrix element access

* Remove wait time calculations

Since wait time is not penalised (arriving early is OK), there is no point in tracking it

* Make ORTEC's dynamic params available from Python

I am unsure if we should use these, but we should test before discarding the idea.

* Simplify penalty management

* Remove unused parameters, move penalty management to GeneticAlgorithm

This is significantly better on the benchmark instances than what we had before, because (I think) the penalty updated were not done very well.